changed names, and add functionality to publicavailability..
