+++ ParkingSytemClient/src/adminGui/AdminLogs.fxml
@@ -5,7 +5,7 @@
 <VBox spacing="25" style="-fx-padding: 30; -fx-background-color: #fdfdfd;" 
       xmlns="http://javafx.com/javafx/8" 
       xmlns:fx="http://javafx.com/fxml"
-      fx:controller="clientgui.admin.AdminLogsController">
+      fx:controller="adminGui.AdminLogsController">
 
     <Label text="🧾 System Logs"
            style="-fx-font-size: 24px; -fx-font-weight: bold; -fx-text-fill: #333;" />
