+++ ParkingSytemClient/src/adminGui/AdminOrdersController.java
@@ -49,7 +49,7 @@ public class AdminOrdersController implements ChatIF {
     @FXML
     private void handleBack() {
         try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/clientgui/admin/AdminMainMenu.fxml"));
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/adminGui/AdminMainMenu.fxml"));
             Parent root = loader.load();
 
             AdminMainMenuController controller = loader.getController();
