+++ ParkingSytemClient/src/adminGui/AdminSubscribersController.java
@@ -25,7 +25,7 @@ public class AdminSubscribersController {
 
     private void handleBack() {
         try {
-            FXMLLoader loader = new FXMLLoader(getClass().getResource("/clientgui/admin/AdminMainMenu.fxml"));
+            FXMLLoader loader = new FXMLLoader(getClass().getResource("/adminGui/AdminMainMenu.fxml"));
             Parent root = loader.load();
 
             AdminMainMenuController controller = loader.getController();
