+++ ParkingSytemClient/src/client/ClientMain.java
@@ -8,13 +8,13 @@ import javafx.stage.Stage;
 
 public class ClientMain extends Application {
   public void start(Stage primaryStage) throws Exception {
-    FXMLLoader loader = new FXMLLoader(ClientMain.class.getResource("/subscriberGui/ConnectView.fxml"));
+    FXMLLoader loader = new FXMLLoader(ClientMain.class.getResource("/client/ConnectView.fxml"));
     Scene scene = new Scene((Parent)loader.load());
     primaryStage.setTitle("Connect to Server");
     primaryStage.setScene(scene);
-    primaryStage.show();
+    primaryStage.show();  
   }
-  
+   
   public static void main(String[] args) {
     launch(args);
   }
