+++ ParkingSytemClient/src/client/ConnectView.fxml
@@ -5,7 +5,7 @@
 <?import javafx.scene.layout.*?>
 
 <AnchorPane xmlns="http://javafx.com/javafx" xmlns:fx="http://javafx.com/fxml"
-            fx:controller="subscriberGui.ConnectViewController"
+            fx:controller="client.ConnectViewController"
             prefWidth="450" prefHeight="250">
 
     <VBox alignment="CENTER" spacing="15.0"
